refactor data_gen ops into qa_gen folder from basic folder
